+++ Cargo.toml
@@ -15,7 +15,7 @@ ctrlc = { version = "3", features = ["termination"] }
 data-encoding = "2.4"
 directories = "5.0"
 futures-util = "0.3"
-giganto-client = { git = "https://github.com/aicers/giganto-client.git", tag = "0.15.1" }
+giganto-client = { git = "https://github.com/aicers/giganto-client.git", tag = "0.15.2" }
 humantime = "2.1"
 humantime-serde = "1"
 libc = "0.2"
